+++ packages/react-native-sdk/src/hooks/index.ts
@@ -5,7 +5,6 @@ export * from './useAndroidKeepCallAliveEffect';
 export * from './useIsIosScreenshareBroadcastStarted';
 export * from './useIsInPiPMode';
 export * from './useAutoEnterPiPEffect';
-export * from './useApplyDefaultMediaStreamSettings';
 export * from './useScreenShareButton';
 export * from './useTrackDimensions';
 export * from './useScreenshot';