+++ sample-apps/react-native/dogfood/android/app/src/main/java/io/getstream/rnvideosample/MainActivity.kt
@@ -17,6 +17,7 @@ class MainActivity : ReactActivity() {
         options.enableMediaProjectionService = true
         // for react-navigation
         super.onCreate(null)
+        StreamVideoReactNative.setupCallActivity(this)
     }
 
     /**