+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.16.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.15.1...@stream-io/video-react-native-sdk-1.16.0) (2025-06-03)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.23.4`
+- `@stream-io/video-react-bindings` updated to version `1.6.7`
+
+### Features
+
+- remove camera management hooks on react native ([#1711](https://github.com/GetStream/stream-video-js/issues/1711)) ([644e238](https://github.com/GetStream/stream-video-js/commit/644e238d479397ca37bcb6c1bd1284988475f9f6))
+
+## [1.15.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.15.0...@stream-io/video-react-native-sdk-1.15.1) (2025-06-02)
+
+### Dependency Updates
+
+- `@stream-io/noise-cancellation-react-native` updated to version `0.1.1`
+- `@stream-io/video-client` updated to version `1.23.3`
+- `@stream-io/video-react-bindings` updated to version `1.6.6`
+
+### Bug Fixes
+
+- expo plugin should add foreground permissions on ringing config ([cba9746](https://github.com/GetStream/stream-video-js/commit/cba97469966aa35b67f380d82272a9bf3a8e2ef2))
+- report to callkit that we do not support holding ([#1809](https://github.com/GetStream/stream-video-js/issues/1809)) ([e779d3f](https://github.com/GetStream/stream-video-js/commit/e779d3fbdda30be5db5c42ee4655d8311dd53d60))
+
+## [1.15.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.14.2...@stream-io/video-react-native-sdk-1.15.0) (2025-05-27)
+
+### Features
+
+- **android:** Accept incoming call without device unlock ([#1806](https://github.com/GetStream/stream-video-js/issues/1806)) ([6b8ee36](https://github.com/GetStream/stream-video-js/commit/6b8ee36323c0c352742e23bf845eed47c581e6ab))
+
 ## [1.14.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-1.14.1...@stream-io/video-react-native-sdk-1.14.2) (2025-05-22)
 
 ### Dependency Updates