+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,35 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.18.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.18.2...@stream-io/video-react-sdk-1.18.3) (2025-06-03)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.23.4`
+- `@stream-io/video-react-bindings` updated to version `1.6.7`
+
+## [1.18.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.18.1...@stream-io/video-react-sdk-1.18.2) (2025-06-02)
+
+### Dependency Updates
+
+- `@stream-io/audio-filters-web` updated to version `0.4.2`
+- `@stream-io/video-client` updated to version `1.23.3`
+- `@stream-io/video-react-bindings` updated to version `1.6.6`
+
+### Bug Fixes
+
+- use AudioContext for Safari ([#1810](https://github.com/GetStream/stream-video-js/issues/1810)) ([63542f4](https://github.com/GetStream/stream-video-js/commit/63542f419efa475c7acf50f053621ace74a1eff4))
+
+## [1.18.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.18.0...@stream-io/video-react-sdk-1.18.1) (2025-05-26)
+
+### Dependency Updates
+
+- `@stream-io/audio-filters-web` updated to version `0.4.1`
+
+### Bug Fixes
+
+- add audio context state tracing ([#1805](https://github.com/GetStream/stream-video-js/issues/1805)) ([0d86623](https://github.com/GetStream/stream-video-js/commit/0d8662398a2e8b5c07bf3ef5b68faf0a4003c702))
+
 ## [1.18.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.17.1...@stream-io/video-react-sdk-1.18.0) (2025-05-22)
 
 ### Dependency Updates